+++ src/main/java/com/tie/repository/DealRepository.java
@@ -11,6 +11,6 @@ import com.tie.model.Deal;
  * @author Ahmed El-Deeb
  *
  */
-public interface DealRepository extends JpaRepository<Deal, Long> {
+public interface DealRepository extends JpaRepository<Deal, Integer> {
 
 }
